Ui [viewmodel, views] changes

+++ app/src/main/java/br/goldbach/activities/ui/adapter/ActivitiesAdapter.kt
@@ -1,11 +1,15 @@
 package br.goldbach.activities.ui.adapter
 
+import android.app.AlertDialog
+import android.content.Context
+import android.content.DialogInterface
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.compose.ui.window.Dialog
+import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.recyclerview.widget.RecyclerView
 import br.goldbach.activities.R
@@ -17,6 +21,7 @@ import br.goldbach.activities.ui.viewmodel.HomeViewModel
 class ActivitiesAdapter(
     private val homeViewModel: HomeViewModel,
     private val fragmentManager: FragmentManager,
+    private val context: Context,
     private var activities: List<Activity> = listOf(),
     private var fragmentNumber: Int = 1
 ) : RecyclerView.Adapter<ActivitiesAdapter.ViewHolder>() {
@@ -41,20 +46,38 @@ class ActivitiesAdapter(
         holder.description.text = activity.description
 
         holder.deleteBtn.setOnClickListener {
-            homeViewModel.deleteActivity(activity)
+            AlertDialog.Builder(context)
+                .setTitle("Delete Activity")
+                .setMessage("Do you really want to delete this activity?")
+                .setPositiveButton("Yes") { _, _ ->
+                    homeViewModel.deleteActivity(activity)
+                }
+                .setNegativeButton("Cancel") { _, _ ->
+                    return@setNegativeButton
+                }
+                .create()
+                .show()
         }
 
-        if(fragmentNumber != 3) {
+        if (fragmentNumber != 3) {
             holder.changeStatusBtn?.setOnClickListener {
-                activity.status = when(fragmentNumber) {
+                activity.status = when (fragmentNumber) {
                     1 -> ActivityStatus.ONGOING
                     else -> ActivityStatus.DONE
                 }
-                homeViewModel.changeActivityStatus(activity)
+                AlertDialog.Builder(context)
+                    .setTitle("Change Status")
+                    .setMessage("Do you really want to change the status of this activity?")
+                    .setPositiveButton("Yes") { _, _ ->
+                        homeViewModel.changeActivityStatus(activity)
+                    }
+                    .setNegativeButton("Cancel") { _, _ ->
+                        return@setNegativeButton
+                    }
+                    .create()
+                    .show()
             }
         }
-
-
     }
 
     override fun getItemCount(): Int {
